+++ Common/Entities/Product.cs
@@ -3,9 +3,9 @@
     public class Product
     {
 
-        public long seller_id { get; set; }
+        public int seller_id { get; set; }
 
-        public long product_id { get; set; }
+        public int product_id { get; set; }
 
         public string name { get; set; } = "";
 
@@ -15,9 +15,9 @@
 
         public string description { get; set; } = "";
 
-        public decimal price { get; set; }
+        public float price { get; set; }
 
-        public decimal freight_value { get; set; }
+        public float freight_value { get; set; }
 
         public string status { get; set; } = "approved";
